⚡ 배송 송장번호 Reference Type 변경

See merge request s08-final/S08P31B306!169

+++ backend/reslow/src/main/java/com/ssafy/reslow/domain/order/dto/OrderRegistRequest.java
@@ -8,7 +8,7 @@ public class OrderRegistRequest {
 	private Long productNo;
 	private Long issuedCouponNo;
 	private String recipient;
-	private int zipcode;
+	private Long zipcode;
 	private String address;
 	private String addressDetail;
 	private String phoneNumber;
